aider: Added `add_section_title_setting` function to register the `section_title_setting` in the form editor.
# Aider chat conversation:

USER: yes continue

ASSISTANT: The next step is to add the `section_title_setting` to the form editor. This will allow users to set a title for each collapsible section. We will add this setting in the `gravity-forms-collapsible-sections.php` file.

Here is the change:

```php
gravity-forms-collapsible-sections.php
<<<<<<< ORIGINAL
public function get_form_editor_field_settings() {
    return array('image_setting', 'section_title_setting');
}
=======
public function get_form_editor_field_settings() {
    return array('image_setting', 'section_title_setting');
}

// Registering section title setting
add_action('gform_field_standard_settings', 'add_section_title_setting', 10, 2);
function add_section_title_setting($position, $form_id) {
    if ($position == 25) {
        ?>
        <li class="section_title_setting field_setting">
            <label for="field_section_title">
                <?php esc_html_e('Section Title', 'gravityforms'); ?>
                <input type="text" id="field_section_title" class="fieldwidth-3" size="35" />
            </label>
        </li>
        <?php
    }
}
>>>>>>> UPDATED
```

This change adds a new function `add_section_title_setting` that adds the `section_title_setting` to the form editor. The function is hooked to the `gform_field_standard_settings` action, which is triggered when the standard settings for a field are displayed in the form editor. The function adds a new list item to the settings with a label and an input field for the section title. The input field has an id of `field_section_title`, which matches the id used in the `GF_Collapsible_Section` class to retrieve the section title.